+++ client/src/commands/code_analysis.ts
@@ -119,6 +119,42 @@ let resultsToDiagnostics = (
           }
         }
       }
+
+      if (item.message.endsWith(" is never used")) {
+        {
+          let codeAction = new CodeAction("Remove unused");
+          codeAction.kind = CodeActionKind.RefactorRewrite;
+
+          let codeActionEdit = new WorkspaceEdit();
+
+          // In the future, it would be cool to have an additional code action
+          // here for automatically removing whatever the thing that's dead is.
+          codeActionEdit.replace(
+            Uri.parse(item.file),
+            // Make sure the full line is replaced
+
+            new Range(
+              new Position(item.range[0], item.range[1]),
+              new Position(item.range[2], item.range[3])
+            ),
+            // reanalyze seems to add two extra spaces at the start of the line
+            // content to replace.
+            ""
+          );
+
+          codeAction.edit = codeActionEdit;
+
+          if (diagnosticsResultCodeActions.has(item.file)) {
+            diagnosticsResultCodeActions
+              .get(item.file)
+              .push({ range: issueLocationRange, codeAction });
+          } else {
+            diagnosticsResultCodeActions.set(item.file, [
+              { range: issueLocationRange, codeAction },
+            ]);
+          }
+        }
+      }
     }
   });