+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Complete `%todo`. https://github.com/rescript-lang/rescript-vscode/pull/981
 - Add code action for extracting a locally defined module into its own file. https://github.com/rescript-lang/rescript-vscode/pull/983
 - Add code action for expanding catch-all patterns. https://github.com/rescript-lang/rescript-vscode/pull/987
+- Add code actions for removing unused code (per item and for an entire file), driven by `reanalyze`.
 
 ## 1.50.0